+++ PoshBot/Classes/CommandExecutor.ps1
@@ -74,7 +74,9 @@ class CommandExecutor {
         if ($authorized) {
 
             # Add reaction telling the user that the command is being executed
-            $this._bot.Backend.AddReaction($Message, [ReactionType]::Processing)
+            if ($this._bot.Configuration.AddCommandReactions) {
+                $this._bot.Backend.AddReaction($Message, [ReactionType]::Processing)
+            }
 
             if ($cmdExecContext.Command.AsJob) {
                 # Kick off job and add to job tracker
@@ -160,10 +162,18 @@ class CommandExecutor {
                         # Determine if job had any terminating errors
                         if ($cmdExecContext.Result.Streams.Error.Count -gt 0) {
                             $cmdExecContext.Result.Success = $false
-                            $this._bot.Backend.AddReaction($cmdExecContext.Message, [ReactionType]::Failure)
                         } else {
                             $cmdExecContext.Result.Success = $true
-                            $this._bot.Backend.AddReaction($cmdExecContext.Message, [ReactionType]::Success)
+                        }
+
+                        # Send a success or fail reaction
+                        if ($this._bot.Configuration.AddCommandReactions) {
+                            if ($cmdExecContext.Result.Success) {
+                                $reaction = [ReactionType]::Success
+                            } else {
+                                $reaction = [ReactionType]::Failure
+                            }
+                            $this._bot.Backend.AddReaction($cmdExecContext.Message, $reaction)
                         }
 
                         Write-Debug -Message "[CommandExecutor:ReceiveJob] Job results:`n$($cmdExecContext.Result | ConvertTo-Json)"
@@ -180,7 +190,9 @@ class CommandExecutor {
                 $this._jobTracker.Remove($cmdExecContext.Id)
 
                 # Remove the reaction specifying the command is in process
-                $this._bot.Backend.RemoveReaction($cmdExecContext.Message, [ReactionType]::Processing)
+                if ($this._bot.Configuration.AddCommandReactions) {
+                    $this._bot.Backend.RemoveReaction($cmdExecContext.Message, [ReactionType]::Processing)
+                }
 
                 # Track number of commands executed
                 if ($cmdExecContext.Result.Success) {