+++ PoshBot/Public/Get-PoshBotConfiguration.ps1
@@ -26,6 +26,7 @@ function Get-PoshBotConfiguration {
         AlternateCommandPrefixSeperators : {:, ,, ;}
         SendCommandResponseToPrivate     : {}
         MuteUnknownCommand               : False
+        AddCommandReactions              : True
 
         Gets the bot configuration located at [C:\Users\joeuser\.poshbot\Cherry2000.psd1].
     .EXAMPLE